Simplify files before going further
Fixes #1

Consolidate related logic into fewer files to simplify the codebase.

* **Plugins.cs**: Add new file merging contents of `ClearwaterPlugin.cs`, `KustoPlugin.cs`, and `LargeRequestHelper.cs`. Adjust namespaces and class names to avoid conflicts.
* **Utilities.cs**: Add new file merging contents of `KustoHelper.cs` and `Logger.cs`. Adjust namespaces and class names to avoid conflicts.
* **ClearwaterChatService.cs**: Update references to `ClearwaterPlugin` and `KustoPlugin` to use the new `Plugins` class.
* **Program.cs**: Update references to `ClearwaterPlugin` and `KustoPlugin` to use the new `Plugins` class.
* **Remove files**: Delete `ClearwaterPlugin.cs`, `KustoPlugin.cs`, `LargeRequestHelper.cs`, `KustoHelper.cs`, and `Logger.cs`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/dciborow/azure-openai-kusto-chat/issues/1?shareId=XXXX-XXXX-XXXX-XXXX).
