+++ src/ClassLibrary1/Plugins.cs
@@ -1,10 +1,76 @@
-﻿namespace Microsoft.AzureCore.ReadyToDeploy.Vira
+namespace Microsoft.AzureCore.ReadyToDeploy.Vira
 {
+    using System;
     using System.Collections.Generic;
     using System.ComponentModel;
     using System.Text.Json;
-
+    using System.Threading.Tasks;
     using global::Microsoft.SemanticKernel;
+
+    public class ClearwaterPlugin
+    {
+        /// <summary>
+        /// Retrieves build information by organization name and build ID.
+        /// </summary>
+        [KernelFunction("get_build_info")]
+        [Description("Retrieves build information by organization name and build ID.")]
+        [return: Description("Returns the build information as a JSON string or a message if no build is found.")]
+        public async Task<string> GetBuildInfoAsync(
+            [Description("The name of the organization (e.g., 'msazure').")] string orgName,
+            [Description("The build ID to look up.")] string buildId)
+            => await ExecuteAndLogQueryAsync("Build", orgName, buildId, "No build found");
+
+        /// <summary>
+        /// Retrieves work items linked to a specific build.
+        /// </summary>
+        [KernelFunction("get_workitem_by_org_build")]
+        [Description("Retrieves work items linked to a specific build.")]
+        [return: Description("Returns the work items associated with the build in JSON format or a message if no work items are found.")]
+        public async Task<string> GetWorkItemsByBuildAsync(
+            [Description("The name of the organization (e.g., 'msazure').")] string orgName,
+            [Description("The build ID to look up.")] string buildId)
+            => await ExecuteAndLogQueryAsync("BuildWorkItem", orgName, buildId, "No work items found");
+
+        /// <summary>
+        /// Retrieves commits linked to a specific build.
+        /// </summary>
+        [KernelFunction("get_commits_by_org_build")]
+        [Description("Retrieves commits linked to a specific build.")]
+        [return: Description("Returns the commits associated with the build in JSON format or a message if no commits are found.")]
+        public async Task<string> GetCommitsByBuildAsync(
+            [Description("The name of the organization (e.g., 'msazure').")] string orgName,
+            [Description("The build ID to look up.")] string buildId)
+            => await ExecuteAndLogQueryAsync("BuildChange", orgName, buildId, "No commits found");
+
+        /// <summary>
+        /// Retrieves Pull Requests linked to a specific build by performing a join between BuildChange and PullRequest.
+        /// </summary>
+        [KernelFunction("get_pull_requests_by_build")]
+        [Description("Retrieves Pull Requests linked to a specific build by performing a join between BuildChange and PullRequest.")]
+        [return: Description("Returns the pull requests associated with the build in JSON format or a message if no pull requests are found.")]
+        public async Task<string> GetPullRequestsByBuildAsync(
+            [Description("The name of the organization (e.g., 'msazure').")] string orgName,
+            [Description("The build ID to look up.")] string buildId)
+        {
+            Logger.LogFunctionCall("ClearwaterPlugin.GetPullRequestsByBuildAsync", orgName, buildId);
+
+            var pullRequests = await KustoHelper.RetrievePullRequestsByBuildIdAsync(orgName, buildId);
+            Logger.LogJson("Tool", pullRequests);
+
+            return string.IsNullOrEmpty(pullRequests) ? $"No pull requests found for Org: {orgName} and BuildId: {buildId}" : pullRequests;
+        }
+
+        // Central method to handle execution and logging for simple queries
+        private async Task<string> ExecuteAndLogQueryAsync(string tableName, string orgName, string buildId, string noDataMessage)
+        {
+            Logger.LogFunctionCall($"ClearwaterPlugin.Get{tableName}ByOrgAndBuildAsync", orgName, buildId);
+
+            var result = await KustoHelper.RetrieveDataByOrgAndBuildIdAsync(tableName, orgName, buildId);
+            Logger.LogJson("Tool", result);
+
+            return string.IsNullOrEmpty(result) ? $"{noDataMessage} for Org: {orgName} and BuildId: {buildId}" : result;
+        }
+    }
 
     public class KustoPlugin
     {
@@ -106,7 +172,6 @@
             }
         }
 
-
         /// <summary>
         /// Provides a list of tables in the specified Kusto database.
         /// </summary>
@@ -160,6 +225,74 @@
             Uri = uri;
             DatabaseName = databaseName;
             Description = description;
+        }
+    }
+
+    public class LargeRequestHelper
+    {
+        private readonly int MaxTokenSize = 100000;  // Maximum token size per GPT request
+        private readonly int MaxRowsPerRequest = 500;  // Maximum rows to send in one request
+        private string _fullOutput = string.Empty;  // Buffer to store the final output
+
+        /// <summary>
+        /// Processes a large data set in batches and interacts with GPT to determine how much more can be processed.
+        /// </summary>
+        /// <param name="query">The Kusto query to execute.</param>
+        /// <param name="clusterUri">The URI of the Kusto cluster.</param>
+        /// <param name="databaseName">The Kusto database name.</param>
+        /// <returns>Returns the final processed output as a string.</returns>
+        public async Task<string> ProcessLargeDataSetAsync(string query, string clusterUri, string databaseName)
+        {
+            int pageSize = MaxRowsPerRequest;
+            int pageIndex = 0;
+            bool continueProcessing = true;
+
+            while (continueProcessing)
+            {
+                // Fetch the next batch of data
+                string batchResult = await KustoHelper.ExecuteKustoQueryForClusterAsync(clusterUri, databaseName, query, true, pageSize, pageIndex);
+
+                if (string.IsNullOrEmpty(batchResult))
+                    break;  // No more data, end the loop
+
+                // Append to the full output
+                _fullOutput += batchResult;
+
+                // Check if GPT can handle more data
+                bool canProcessMore = AskGPTIfMoreDataCanBeProcessed(_fullOutput);
+
+                if (!canProcessMore)
+                {
+                    // Ask user if they want to continue processing more chunks
+                    continueProcessing = AskUserIfTheyWantToContinue();
+                }
+                else
+                {
+                    // Increase page index to fetch the next batch
+                    pageIndex++;
+                }
+            }
+
+            return _fullOutput;
+        }
+
+        /// <summary>
+        /// Asks GPT whether it can handle more rows based on the current data size.
+        /// </summary>
+        private bool AskGPTIfMoreDataCanBeProcessed(string currentOutput) =>
+            // Here, send the currentOutput to GPT and ask if it can handle more data.
+            // For simplicity, let's simulate GPT's decision-making:
+            currentOutput.Length < MaxTokenSize;
+
+        /// <summary>
+        /// Asks the user if they want to continue processing more chunks.
+        /// </summary>
+        private bool AskUserIfTheyWantToContinue()
+        {
+            // Notify the user about the current batch, chunks processed, and ask if they want to continue
+            Console.WriteLine($"Processed {_fullOutput.Length} tokens so far. Do you want to process more? (y/n)");
+            string input = Console.ReadLine();
+            return input?.ToLower() == "y";
         }
     }
 }
